Split the managed_switch firmware into apps and device drivers
Each device now has one driver in firmware/src/drivers/: Terminal (UART0),
Leds (GPIO), Clock (CLINT, with Ticker), Dp83867 and Switch. Switch wraps
the four SatCat5 ConfigBus devices (switch core, CPU port, port statistics
and MDIO) as one switching hub on top of satcat5-pac, replacing the
hand-written register offsets. The console keeps only line editing and
the commands.

The release .text shrinks from 33572 to 33524 bytes.

+++ projects/managed_switch/firmware/src/drivers/dp83867/registers.rs
@@ -1,10 +1,10 @@
 //! DP83867 のレジスタの番号とビット。
 
-pub const BMCR: u32 = 0x00;
-pub const CFG1: u32 = 0x09;
-pub const REGCR: u32 = 0x0D;
-pub const ADDAR: u32 = 0x0E;
-pub const PHYSTS: u32 = 0x11;
+pub const BMCR: u8 = 0x00;
+pub const CFG1: u8 = 0x09;
+pub const REGCR: u8 = 0x0D;
+pub const ADDAR: u8 = 0x0E;
+pub const PHYSTS: u8 = 0x11;
 
 // 拡張レジスタは MDIO で直接指せないため、番号を ADDAR に書いて指す。
 pub const RGMIICTL: u16 = 0x0032;
